ispravljeno za exit

+++ src/app/components/store-check-page/store-check-page.component.ts
@@ -129,20 +129,24 @@ export class StoreCheckPageComponent implements OnInit {
   }
 
   public exit() {
-    const dialogRef = this.dialog.open(AreYouSureDialogComponent);
-    dialogRef.afterClosed()
-      .subscribe(res => {
-        console.log(res)
-        if (res) {
-          this.router.navigate(['/chooseObject/' + this.workModel]);
-          if (!this.resolveFeedbacks) {
-            this.objectStoreCheckService.deleteUnfinishedObjectStoreCheck("ppetrovic").subscribe(data => {
-              console.log(data);
-            });
+    if (this.showFinishButton) {
+      const dialogRef = this.dialog.open(AreYouSureDialogComponent);
+      dialogRef.afterClosed()
+        .subscribe(res => {
+          console.log(res)
+          if (res) {
+            this.router.navigate(['/chooseObject/' + this.workModel]);
+            if (!this.resolveFeedbacks) {
+              this.objectStoreCheckService.deleteUnfinishedObjectStoreCheck("ppetrovic").subscribe(data => {
+                console.log(data);
+              });
+            }
           }
         }
-      }
-      )
+        )
+    } else {
+      this.router.navigate(['/chooseObject/' + this.workModel]);
+    }
   }
 
 }
